Mark IDisplayService as deprecated.
Nothing uses this in the Android tree that I can find, but for being
cautious and according to Treble rules, deprecating this service.

Funny story, I implemented this service around midnight or so because I
couldn't wait to get that vsync data out of there!

If any device is using this service and needs to launch on Android T,
please reach out. We could deliver another solution or restore this
temporarily. However, in general, using binder calls to do frame timing
is pretty errorprone. You may prefer sync fences or similar.

Fixes: 141930622
Test: boot cf, lshal | grep display
Test: atest vts_treble_vintf_{vendor,framework}_test
Change-Id: Id7fcf18c74aedc8f15af4529552b2138d545224a